Move both keybase_cmd and keybase_error to their own module
