+++ src/keybase/error.rs
@@ -1,0 +1,7 @@
+error_chain! {
+	foreign_links {
+			Parsing(::serde_json::error::Error);
+			IOErr(::std::io::Error);
+			UTF8Err(std::string::FromUtf8Error);
+	}
+}
